+++ app/dtos/spot_models.py
@@ -1,31 +1,30 @@
 from pydantic import BaseModel, Field
 from datetime import datetime
+from typing import Optional
 
 
 class spot_pydantic(BaseModel):
     kor_name: str = Field(max_length=255)
-    eng_name: str = Field(default=None, max_length=255)
+    eng_name: Optional[str] = Field(default=None, max_length=255)
     address: str = Field(max_length=255)
-    zip: str = Field(max_length=10)
-    url: str = Field(default=None, max_length=2083)
+
+    url: Optional[str] = Field(default=None, max_length=2083)
     image_url: str = Field(max_length=2083)
     map_url: str = Field(max_length=2083)
-    likes: int = None
-    satisfaction: float = None
+
     spot_category: int
-    phone_number: str = Field(default=None, max_length=300)
-    business_status: bool = None
-    business_hours: str = Field(default=None, max_length=255)
+    phone_number: Optional[str] = Field(default=None, max_length=300)  # Optional로 변경
+    business_status: Optional[bool] = None
+    business_hours: Optional[str] = Field(default=None, max_length=255)
     order: int
     day_x: int
-    spot_time: str = None
-    latitude: float = None  
-    longitude: float = None  
-    distance_from_prev: float = None  
+    spot_time: Optional[str] = None
+    latitude: Optional[float] = None
+    longitude: Optional[float] = None
+
 
 class spots_pydantic(BaseModel):
     spots: list[spot_pydantic]
-
 
 
 def calculate_trip_days(start_date_str: str, end_date_str: str) -> int:
